LLM/pyproject.toml
[tool.poetry]
name = "llm"
version = "0.1.0"
description = ""
authors = ["dotpep <pepelnine@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "2.31.0"
python-dotenv = "1.0.0"
transformers = "4.44.2"
huggingface-hub = "^0.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
